Files divided into packages, fix 404 error
